corrected button formatting
changed the button formatting for the socials and download resume buttons, also added the button style from the main index file into the indiviual project pages. Just need to make a preloader and ensure that when the project pages are loaded, the video will actually play on the first time.